hw5: fix bugs in external sort

+++ hw5/src/external_sort.h
@@ -6,6 +6,6 @@
 #include <string>
 #include <iostream>
 
-void externalSort(std::vector<IO_Helper *> helperVec, int bufferSize);
+void externalSort(std::vector<IO_Helper *> helperVec, long bufferSize);
 
 #endif
